pkg/generator: Refactor generator_test.go
