+++ pkg/generator/gen_api_register_test.go
@@ -1,0 +1,68 @@
+// Copyright 2018 The Operator-SDK Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package generator
+
+import (
+	"bytes"
+	"testing"
+)
+
+const registerExp = `package v1alpha1
+
+import (
+	sdkK8sutil "github.com/operator-framework/operator-sdk/pkg/util/k8sutil"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+)
+
+const (
+	version   = "v1alpha1"
+	groupName = "app.example.com"
+)
+
+var (
+	SchemeBuilder = runtime.NewSchemeBuilder(addKnownTypes)
+	AddToScheme   = SchemeBuilder.AddToScheme
+	// SchemeGroupVersion is the group version used to register these objects.
+	SchemeGroupVersion = schema.GroupVersion{Group: groupName, Version: version}
+)
+
+func init() {
+	sdkK8sutil.AddToSDKScheme(AddToScheme)
+}
+
+// addKnownTypes adds the set of types defined in this package to the supplied scheme.
+func addKnownTypes(scheme *runtime.Scheme) error {
+	scheme.AddKnownTypes(SchemeGroupVersion,
+		&AppService{},
+		&AppServiceList{},
+	)
+	metav1.AddToGroupVersion(scheme, SchemeGroupVersion)
+	return nil
+}
+`
+
+func TestGenRegister(t *testing.T) {
+	buf := &bytes.Buffer{}
+	if err := renderAPIRegisterFile(buf, appKind, appGroupName, appVersion); err != nil {
+		t.Error(err)
+		return
+	}
+	if registerExp != buf.String() {
+		t.Errorf("want %v, got %v", registerExp, buf.String())
+	}
+}
