+++ pkg/generator/gen_deps_test.go
@@ -1,0 +1,41 @@
+// Copyright 2018 The Operator-SDK Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package generator
+
+import (
+	"bytes"
+	"testing"
+)
+
+func TestGenGopkg(t *testing.T) {
+	buf := &bytes.Buffer{}
+	if err := renderGopkgTomlFile(buf); err != nil {
+		t.Error(err)
+		return
+	}
+
+	if gopkgTomlTmpl != buf.String() {
+		t.Errorf("want %v, got %v", gopkgTomlTmpl, buf.String())
+	}
+
+	buf = &bytes.Buffer{}
+	if err := renderGopkgLockFile(buf); err != nil {
+		t.Error(err)
+		return
+	}
+	if gopkgLockTmpl != buf.String() {
+		t.Errorf("want %v, got %v", gopkgLockTmpl, buf.String())
+	}
+}
