+++ pkg/generator/gen_handler_test.go
@@ -1,0 +1,101 @@
+// Copyright 2018 The Operator-SDK Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package generator
+
+import (
+	"bytes"
+	"testing"
+)
+
+const handlerExp = `package stub
+
+import (
+	"context"
+
+	"github.com/example-inc/app-operator/pkg/apis/app/v1alpha1"
+
+	"github.com/operator-framework/operator-sdk/pkg/sdk"
+	"github.com/sirupsen/logrus"
+	"k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+)
+
+func NewHandler() sdk.Handler {
+	return &Handler{}
+}
+
+type Handler struct {
+	// Fill me
+}
+
+func (h *Handler) Handle(ctx context.Context, event sdk.Event) error {
+	switch o := event.Object.(type) {
+	case *v1alpha1.AppService:
+		err := sdk.Create(newbusyBoxPod(o))
+		if err != nil && !errors.IsAlreadyExists(err) {
+			logrus.Errorf("Failed to create busybox pod : %v", err)
+			return err
+		}
+	}
+	return nil
+}
+
+// newbusyBoxPod demonstrates how to create a busybox pod
+func newbusyBoxPod(cr *v1alpha1.AppService) *v1.Pod {
+	labels := map[string]string{
+		"app": "busy-box",
+	}
+	return &v1.Pod{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Pod",
+			APIVersion: "v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "busy-box",
+			Namespace: cr.Namespace,
+			OwnerReferences: []metav1.OwnerReference{
+				*metav1.NewControllerRef(cr, schema.GroupVersionKind{
+					Group:   v1alpha1.SchemeGroupVersion.Group,
+					Version: v1alpha1.SchemeGroupVersion.Version,
+					Kind:    "AppService",
+				}),
+			},
+			Labels: labels,
+		},
+		Spec: v1.PodSpec{
+			Containers: []v1.Container{
+				{
+					Name:    "busybox",
+					Image:   "busybox",
+					Command: []string{"sleep", "3600"},
+				},
+			},
+		},
+	}
+}
+`
+
+func TestGenHandler(t *testing.T) {
+	buf := &bytes.Buffer{}
+	if err := renderHandlerFile(buf, appRepoPath, appKind, appApiDirName, appVersion); err != nil {
+		t.Error(err)
+		return
+	}
+	if handlerExp != buf.String() {
+		t.Errorf("want %v, got %v", handlerExp, buf.String())
+	}
+}
